# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Novel Gen API",
    "description": "AI Long-form Novel Generation System",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/projects": {
      "post": {
        "tags": [
          "项目"
        ],
        "summary": "Create Project",
        "description": "创建新项目",
        "operationId": "create_project_api_projects_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "项目"
        ],
        "summary": "List Projects",
        "description": "获取用户的所有项目",
        "operationId": "list_projects_api_projects_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectResponse"
                  },
                  "title": "Response List Projects Api Projects Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/projects/{project_id}/worldbuilding": {
      "post": {
        "tags": [
          "世界观"
        ],
        "summary": "Trigger Worldbuilding",
        "description": "触发世界观生成任务（流式）",
        "operationId": "trigger_worldbuilding_api_projects__project_id__worldbuilding_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorldbuildingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorldbuildingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "世界观"
        ],
        "summary": "Get Worldbuilding",
        "description": "获取世界观结果",
        "operationId": "get_worldbuilding_api_projects__project_id__worldbuilding_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorldbuildingResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/worldbuilding/stream/{task_id}": {
      "get": {
        "tags": [
          "世界观"
        ],
        "summary": "Worldbuilding Stream",
        "description": "SSE 流：订阅世界观生成任务的实时 token",
        "operationId": "worldbuilding_stream_api_projects_worldbuilding_stream__task_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/tasks/{task_id}": {
      "get": {
        "tags": [
          "世界观"
        ],
        "summary": "Get Task Status",
        "description": "获取任务状态（用于非流式场景的兜底轮询）",
        "operationId": "get_task_status_api_projects_tasks__task_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/compliance/check": {
      "post": {
        "tags": [
          "世界观"
        ],
        "summary": "Compliance Check",
        "description": "扫描输入文本中的真实国家/城市/品牌，返回 issues 列表。\n\n用于前端在用户提交 core_idea 时给出即时反馈。",
        "operationId": "compliance_check_api_projects_compliance_check_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComplianceCheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplianceCheckResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/tasks/{task_id}": {
      "get": {
        "tags": [
          "任务"
        ],
        "summary": "Get Task Status",
        "operationId": "get_task_status_api_tasks__task_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ComplianceCheckRequest": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "ComplianceCheckRequest"
      },
      "ComplianceCheckResponse": {
        "properties": {
          "compliant": {
            "type": "boolean",
            "title": "Compliant"
          },
          "issues": {
            "items": {},
            "type": "array",
            "title": "Issues"
          }
        },
        "type": "object",
        "required": [
          "compliant",
          "issues"
        ],
        "title": "ComplianceCheckResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProjectCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title"
          },
          "core_idea": {
            "type": "string",
            "minLength": 10,
            "title": "Core Idea"
          },
          "genre": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Genre"
          },
          "tone_style": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tone Style"
          },
          "target_word_count": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000000.0,
                "minimum": 1000.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Word Count",
            "default": 100000
          },
          "target_chapter_count": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 2000.0,
                "minimum": 10.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Chapter Count",
            "default": 90
          }
        },
        "type": "object",
        "required": [
          "title",
          "core_idea"
        ],
        "title": "ProjectCreate"
      },
      "ProjectResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "core_idea": {
            "type": "string",
            "title": "Core Idea"
          },
          "genre": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Genre"
          },
          "tone_style": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tone Style"
          },
          "target_word_count": {
            "type": "integer",
            "title": "Target Word Count"
          },
          "target_chapter_count": {
            "type": "integer",
            "title": "Target Chapter Count"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "core_idea",
          "genre",
          "tone_style",
          "target_word_count",
          "target_chapter_count",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "ProjectResponse"
      },
...
      "TaskStatus": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "status"
        ],
        "title": "TaskStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorldbuildingRequest": {
        "properties": {
          "regenerate": {
            "type": "boolean",
            "title": "Regenerate",
            "description": "是否重新生成",
            "default": false
          }
        },
        "type": "object",
        "title": "WorldbuildingRequest"
      },
      "WorldbuildingResponse": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "status"
        ],
        "title": "WorldbuildingResponse"
      },
      "WorldbuildingResult": {
        "properties": {
          "setting_document": {
            "type": "string",
            "title": "Setting Document"
          },
          "constraints": {
            "additionalProperties": true,
            "type": "object",
            "title": "Constraints"
          },
          "conflict_seeds": {
            "items": {},
            "type": "array",
            "title": "Conflict Seeds"
          }
        },
        "type": "object",
        "required": [
          "setting_document",
          "constraints",
          "conflict_seeds"
        ],
        "title": "WorldbuildingResult"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}